pipeline/requirements.txt
# Data providers
yfinance>=0.2.36
httpx>=0.27.0
feedparser>=6.0.11

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Sentiment analysis
transformers>=4.36.0
torch>=2.1.0

# Configuration
python-dotenv>=1.0.0

# Logging
structlog>=24.1.0
